First pass at correctly using the Side Parameter, ie. it specifies which channel to use for stereo samples. Previously I thought it was set for the Stereo/Right Channel Mode (which seemed redundant). Also bump the version number.

+++ Source/GUI/Assimil8or/Editor/Waveform/WaveformDisplay.h
@@ -34,7 +34,7 @@ private:
     juce::int64 loopLength { 0 };
     int halfHeight { 0 };
     int numPixels { 0 };
-    int samplesPerPixel { 0 };
+    float samplesPerPixel { 0.f };
     int markerStartY { 1 };
     int markerEndY { 0 };
     juce::Rectangle<int> sampleStartHandle;
